added search suggestions based on recorded queries. To make this possible, it was necessary - to create a new search index 'queries' - record all queries made by the user - provide a new search api at /api/suggest.json The code which came out appeared very messy because of a lot of duplications. To clean up, most part of loklak was refactored. Index entry factories for the different index types had been introduced. This feature cares about privacy. No user ip is recorded for queries and no user-related information is stored. Misuse of the recording function is already covered by the DoS protection.
To use the suggest interface, it is possible to make a fuzzy search on
queries. For localhost access it is also possible to make empty searches
which lists all queries in the query database. A large number of
different sort order options are possible. It is possible to order by
first query time, latest query time, number of times a user submitted a
query, number of tweets per day estimated by the latest query result and
much more. Examples:
http://localhost:9100/api/suggest.json?q=soj&orderby=query_count
http://localhost:9100/api/suggest.json?count=20&order=asc
http://localhost:9100/api/suggest.json?count=20&order=desc
http://localhost:9100/api/suggest.json?count=20&orderby=messages_per_day&order=desc