Changed small comment errors in config.properties
In my earlier PR #553, I had earlier used authtype, I changed the name to encryption for it to be clearer. Also removed ssl and added starttls in the code. I did the changes in code but forgot to change it in the comments. Trivial fix (non-code), so merging this to master.