allow for logging to stdout while still daemonizing
In the systemd service unit we want the daemon be of type
forking, so that systemd can reliable determine whether startup
has completed.

We add a setting SS_LOG_TO_STDOUT (cmd line option -o) to start.sh
that triggers this.

For the systemd unit we set the above env variable to 1, set
the type to forking, and in addition specify for stability the
location of the PID file.